Partial error reporting to logs
