+++ apis/Google.Cloud.Diagnostics.Common/Google.Cloud.Diagnostics.Common/ErrorReporting/ErrorReportingOptions.cs
@@ -1,0 +1,34 @@
+﻿// Copyright 2017 Google Inc. All Rights Reserved.
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Google.Cloud.ErrorReporting.V1Beta1;
+using Google.Cloud.Logging.V2;
+
+namespace Google.Cloud.Diagnostics.AspNetCore
+{
+    public enum ReportEventsTo
+    {
+        ErrorReporting,
+        Logging
+    }
+
+    public sealed class ErrorReportingOptions
+    {
+
+        internal ErrorReportingOptions(ReportEventsTo reportEventsTo, 
+            ReportErrorsServiceClient errorClient = null, LoggingServiceV2Client logClient = null)
+        {
+        }
+    }
+}
